Added Android Project
Added initial version of the android project. Adding new contacts and
calling existing contacts in the contact data works. The data isn't
encrypted in any way though, and the date can be wonky with the months
(Though I think I got it fixed). Contact data is in the form of
"year;month;day;number" without the quotes. A settings menu is still
needed, as well as all the geolocation bits. We also need to come up
with a way/location to sync the contact data with the web when a new
contact is added or after a set period or time maybe.